requirements.txt
asgiref==3.8.1
boto3==1.35.31
botocore==1.35.31
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
jmespath==1.0.1
kafka-python==2.0.2
Markdown==3.7
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
